pyproject.toml
# PACKAGE

[tool.poetry]
name = "cinema-paris"
version = "0.1.0"
description = ""
packages = [{ include = "app" }]
authors = ["Hông-Lan Botterman <honglan.botterman@gmail.com>"]
readme = "README.md"

# DEPENDENCIES

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = "^0.31.1"
pydantic = "^2.9.2"
jinja2 = "^3.1.4"
pytest = "^8.3.3"
ruff = "^0.6.9"
requests = "^2.32.3"
pyyaml = "^6.0.2"
pytest-mock = "^3.14.0"
httpx = "^0.27.2"
pytest-asyncio = "^0.24.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 79
fix = true
indent-width = 4
target-version = "py311"

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = ["C", "D", "E", "F", "I", "Q", "W"]
ignore = ["C901", "D107", "D203", "D213"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["D100", "D103"]
"__init__.py" = ["D104"]